Big network refactor ready for video

+++ common/src/main/java/dev/imabad/theatrical/client/blockentities/FresnelRenderer.java
@@ -118,7 +118,7 @@ public class FresnelRenderer extends FixtureRenderer<FresnelBlockEntity> {
                     int r = (color >> 16) & 0xFF;
                     int g = (color >> 8) & 0xFF;
                     int b = color & 0xFF;
-                    int a = (int) (((float) ((intensity * 1) / 255f)) * 255);
+                    int a = (int) (((intensity * 1) / 255f) * 255);
                     poseStack.translate(0, -0.25f, 0.05f);
                     Matrix4f m = poseStack.last().pose();
                     Matrix3f normal = poseStack.last().normal();
